requirements (18) (1).txt
python-telegram-bot==22.7
aiohttp==3.13.5
requests==2.33.1
urllib3==2.6.3
APScheduler==3.11.2
